Backend lead navigator ai/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==23.0.0
python-multipart==0.0.9
starlette==0.37.2
SQLAlchemy==2.0.35
alembic==1.13.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic[email]==2.9.2
numpy==2.1.1
pandas==2.2.3
polars==1.9.0
openpyxl==3.1.5
chardet==5.2.0
google-generativeai==0.7.2
scikit-learn==1.5.2
httpx==0.27.2
requests==2.32.3
python-dateutil==2.9.0.post0
typing-extensions==4.12.2
pydantic-settings==2.5.2
onnxruntime==1.20.0
PyJWT==2.8.0